refactor(`Modal`): move `Modal` & `ModalLayout` into `modal` dir
this mirrors the way we are handling modals inside the other features
